calendar tweaks
