+++ app/components/calendar/Calendar.tsx
@@ -43,6 +43,7 @@ import {
   ViewToggleButton,
   MainContainer,
 } from "./style";
+import CustomToolbar from "./CustomToolbar";
 
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import "react-big-calendar/lib/addons/dragAndDrop/styles.css";
@@ -81,66 +82,10 @@ type Props = {
 };
 
 const CalendarScheduler = ({ role, userid }: Props) => {
-  const CustomToolbar: React.FC<any> = (toolbar) => {
-    const goToBack = () => {
-      toolbar.onNavigate("PREV");
-    };
 
-    const goToNext = () => {
-      toolbar.onNavigate("NEXT");
-    };
 
-    const label = () => {
-      const date = toolbar.date;
-      if (toolbar.view === "month") {
-        return moment(date).format("MMMM YYYY");
-      }
-      if (toolbar.view === "week") {
-        const start = moment(date).startOf("week").format("MMM D");
-        const end = moment(date).endOf("week").format("MMM D, YYYY");
-        return `${start} - ${end}`;
-      }
-      if (toolbar.view === "day") {
-        return moment(date).format("dddd, MMMM D, YYYY");
-      }
-      return "";
-    };
 
-    return (
-      <NavigationContainer>
-        <LeftNav>
-          <NavigationButton onClick={goToBack}>
-            <CircleArrowLeft />
-          </NavigationButton>
-          <span style={{ margin: "0 10px" }}>{label()}</span>
-          <NavigationButton onClick={goToNext}>
-            <CircleArrowRight />
-          </NavigationButton>
-        </LeftNav>
-
-        <ViewToggle>
-          <Button
-            onClick={toolbar.onAddEvent}
-            style={{ backgroundColor: "transparent", padding: "0px 12px" }}
-          >
-            <CirclePlus color="#2B5B76" />
-          </Button>
-          <ViewToggleButton
-            onClick={() => toolbar.onView("month")}
-            $active={toolbar.view === "month"}
-          >
-            Month
-          </ViewToggleButton>
-          <ViewToggleButton
-            onClick={() => toolbar.onView("week")}
-            $active={toolbar.view === "week"}
-          >
-            Week
-          </ViewToggleButton>
-        </ViewToggle>
-      </NavigationContainer>
-    );
-  };
+  
   const [events, setEvents] = useState<Event[]>([]);
   const [newEventTitle, setNewEventTitle] = useState("");
   const [newEventDate, setNewEventDate] = useState("");
@@ -182,6 +127,7 @@ const CalendarScheduler = ({ role, userid }: Props) => {
   }, [isDialogOpen]);
 
   const addOrUpdateEvent = () => {
+    
     if (
       newEventTitle &&
       newEventDate &&
@@ -204,12 +150,22 @@ const CalendarScheduler = ({ role, userid }: Props) => {
         createdAt: new Date(),
         updatedAt: new Date(),
       };
-      console.log(userid);
 
-      if (editingEvent) {
+  
+    
+  
+      if (editingEvent )
+        
+        {
+          if (role !== "admin" && userid !== editingEvent.owner) {
+            alert("You do not have permission to edit this event.");
+            return;
+          }
+    
+          
         setEvents(
           events.map((e) => (e.id === editingEvent.id ? eventData : e))
-        );
+        ) 
         updateEvent(eventData);
       } else {
         setEvents([...events, eventData]);
